V1

+++ MapMaven.Core/Utilities/Scoresaber/PPCurveItem.cs
@@ -2,7 +2,7 @@
 {
     public class PPCurveItem
     {
-        public decimal At { get; set; }
-        public decimal Value { get; set; }
+        public double At { get; set; }
+        public double Value { get; set; }
     }
 }
